+++ packages/widgetbook_cli/lib/src/commands/build_push.dart
@@ -19,7 +19,6 @@ class BuildPushCommand extends CliCommand<BuildPushArgs> {
     required super.context,
     this.processManager = const LocalProcessManager(),
     this.fileSystem = const LocalFileSystem(),
-    this.zipEncoder = const ZipEncoder(),
     this.cacheReader = const CacheReader(),
     WidgetbookHttpClient? cloudClient,
     StorageClient? storageClient,
@@ -56,6 +55,11 @@ class BuildPushCommand extends CliCommand<BuildPushArgs> {
         help: 'Full commit SHA',
       )
       ..addOption(
+        'merged-result-commit',
+        help: 'For GitLab Merged Results, '
+            'this commit will be used for commit status.',
+      )
+      ..addOption(
         'actor',
         help: 'Author of the commit',
       );
@@ -65,7 +69,6 @@ class BuildPushCommand extends CliCommand<BuildPushArgs> {
   final StorageClient storageClient;
   final ProcessManager processManager;
   final FileSystem fileSystem;
-  final ZipEncoder zipEncoder;
   final CacheReader cacheReader;
 
   @override
@@ -98,6 +101,8 @@ class BuildPushCommand extends CliCommand<BuildPushArgs> {
         context.providerSha ??
         currentBranch.sha;
 
+    final mergedResultCommit = results['merged-result-commit'] as String?;
+
     final actor = results['actor'] as String? ?? context.user;
     if (actor == null) {
       throw ActorNotFoundException();
@@ -112,6 +117,7 @@ class BuildPushCommand extends CliCommand<BuildPushArgs> {
       apiKey: apiKey,
       branch: branch,
       commit: commit,
+      mergedResultCommit: mergedResultCommit,
       path: path,
       vendor: context.name,
       actor: actor,
@@ -182,6 +188,7 @@ class BuildPushCommand extends CliCommand<BuildPushArgs> {
         actor: args.actor,
         branch: args.branch,
         sha: args.commit,
+        mergedResultSha: args.mergedResultCommit,
         useCases: cache.useCases,
         addonsConfigs: cache.addonsConfigs,
         size: dirSize,